Fixed Various Bugs & Made Improvements
Added Try-Except-Else clauses to ensure that only the appropriate values were accepted and to make the program not break upon receiving the wrong value.

Changed user input for conversion options to a float, instead of an int. It was causing an issue where if you entered a float value to convert, since it was not an integer it would print an error message. It will now accept decimal values.